+++ org-structure-app-backend/src/main/java/com/ppteam/orgstructureserver/dto/OrganizationalUnitsAggregationInfoDTO.java
@@ -1,11 +1,12 @@
 package com.ppteam.orgstructureserver.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrganizationalUnitsAggregationInfoDTO(
         int totalPositionsAmount,
         int employeesAmount,
         int vacanciesAmount,
-        long totalWageFund,
+        BigDecimal totalWageFund,
         List<JobTitleStatisticsDTO> jobTitlesStatistics) {
 }
